backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
bcrypt==4.0.1
slowapi
reportlab
dotenv
passlib
python-jose
pydantic[email]
